some more security slq injections on deleting users

+++ admin/includes/view_all_users.php
@@ -105,13 +105,20 @@ if (isset($_GET['change_to_sub'])) {
 
 if (isset($_GET['delete'])) {
 
-    $the_user_id = $_GET['delete'];
+    if (isset($_SESSION['user_role'])) {
 
-    $query = "DELETE FROM users WHERE user_id = $the_user_id ";
+        if ($_SESSION['user_role'] == 'admin') {
 
-    $delete_query = mysqli_query($connection, $query);
+            $the_user_id = mysqli_escape_string($connection, $_GET['delete']);
 
-    header("Location: users.php");
+            $query = "DELETE FROM users WHERE user_id = $the_user_id ";
+
+            $delete_query = mysqli_query($connection, $query);
+
+            header("Location: users.php");
+        }
+
+    }
 
 }
